refactor(date): affiche la date sélectionnée par défaut
